pyproject.toml
[tool.poetry]
name = "currency-converter"
version = "0.1.0"
description = ""
authors = ["Nikolai <nicolas.gordienko@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "converter"}]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
flask = "^3.0.0"
gunicorn = "^21.2.0"
flake8 = "^6.1.0"
uvicorn = "^0.25.0"
pydantic = "^2.5.3"
fastapi = "^0.108.0"
requests = "^2.31.0"
alembic = "^1.13.1"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
sqlalchemy-utils = "^0.41.1"
responses = "^0.24.1"
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
coverage = "^7.3.4"
pyyaml = "^6.0.1"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
